requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
bcrypt
jinja2
python-multipart
